+++ snowpack/Utils.h
@@ -65,7 +65,7 @@ bool booleanTime(const double& JulianDate, double days_between,
 void deleteOldOutputFiles(const std::string& outdir, const std::string& experiment,
                           const std::string& stationID, const unsigned int& nSlopes);
 
-void averageFluxTimeSeries(const int& n_steps, const bool& useCanopyModel,
+void averageFluxTimeSeries(const size_t& n_steps, const bool& useCanopyModel,
                            SurfaceFluxes& Sdata, SnowStation& Xdata);
 
 void typeToCode(int *F1, int *F2, int *F3, int type);
